refactor: isolate target rejection payload
