+++ crates/openengine-cluster-protocol/src/native_v2_target.rs
@@ -8,6 +8,10 @@ use std::fmt;
 
 use serde::{Deserialize, Serialize};
 
+mod rejection;
+
+pub use rejection::{TargetRunRejection, MAX_TARGET_RUN_REJECTION_MESSAGE_BYTES};
+
 use crate::{
     ConnectionKey, EnvironmentVariableName, NativeV2RunValueError, RunId, RunSubmission,
     TargetRunProfilesDiscovery, MAX_DECLARED_ENVIRONMENT_NAMES,
@@ -28,7 +32,6 @@ pub const GITHUB_APP_INSTALLATION_CONNECTION_KIND: &str = "github-app-installati
 
 const MAX_CONNECTION_VALUE_BYTES: usize = 64 * 1024;
 const MAX_CONNECTION_BYTES: usize = 256 * 1024;
-pub const MAX_TARGET_RUN_REJECTION_MESSAGE_BYTES: usize = 1_024;
 
 #[derive(Clone, Copy, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "snake_case")]
@@ -246,49 +249,6 @@ fn redacted(secret: &Option<String>) -> Option<&'static str> {
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct TargetRunReceipt {
     pub run_id: RunId,
-}
-
-/// Public, secret-free reason that a target rejected a submitted run before admission.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize)]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-pub struct TargetRunRejection {
-    message: String,
-}
-
-impl TargetRunRejection {
-    pub fn new(message: impl Into<String>) -> Result<Self, NativeV2RunValueError> {
-        let message = message.into();
-        if message.is_empty()
-            || message.len() > MAX_TARGET_RUN_REJECTION_MESSAGE_BYTES
-            || message.chars().any(char::is_control)
-        {
-            return Err(NativeV2RunValueError(
-                "target run rejection message must be 1..=1024 non-control UTF-8 bytes",
-            ));
-        }
-        Ok(Self { message })
-    }
-
-    #[must_use]
-    pub fn message(&self) -> &str {
-        &self.message
-    }
-}
-
-impl<'de> Deserialize<'de> for TargetRunRejection {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(deny_unknown_fields, rename_all = "camelCase")]
-        struct Wire {
-            message: String,
-        }
-
-        let wire = Wire::deserialize(deserializer)?;
-        Self::new(wire.message).map_err(serde::de::Error::custom)
-    }
 }
 
 /// A hosted authority requires a run so it can route to one exact task attempt. Direct targets
@@ -520,22 +480,5 @@ mod tests {
         assert!(debug.contains("TOKEN"));
         assert!(!debug.contains("secret-value"));
         assert!(StaticConnectionValues::new(BTreeMap::new()).is_err());
-    }
-
-    #[test]
-    fn target_run_rejection_is_bounded_and_closed() {
-        let rejection = TargetRunRejection::new("required input binding is missing").assert_value();
-        let encoded = serde_json::to_value(&rejection).assert_value();
-        assert_eq!(
-            encoded,
-            serde_json::json!({"message": "required input binding is missing"})
-        );
-        assert!(TargetRunRejection::new("x".repeat(1_025)).is_err());
-        assert!(
-            serde_json::from_value::<TargetRunRejection>(
-                serde_json::json!({"message": "rejected", "extra": true})
-            )
-            .is_err()
-        );
     }
 }
